# ⚡️ Simplify and speed up `mailbox-list` parsing