Note that this fixes `MBX_LIST_FLAGS` and `#mbx_list_flags`, which were previously unused, and uses them too.

This gives a significant performance boost:

```
Warming up --------------------------------------
list_with_various_flag_capitalizations            38.976k i/s
        -     39.270k times in 1.007548s (25.66μs/i)
   rfc3501_7.2.2_LIST_response_example            50.042k i/s
        -     52.294k times in 1.044997s (19.98μs/i)
                  utf8_in_list_mailbox            41.722k i/s
        -     44.550k times in 1.067774s (23.97μs/i)
                           xlist_inbox            43.242k i/s
        -     46.387k times in 1.072725s (23.13μs/i)
Calculating -------------------------------------
                                       v0.4.5-8-g4611404d       0.4.5
list_with_various_flag_capitalizations            39.943k     35.777k i/s
        -    116.927k times in 2.927367s 3.268215s
   rfc3501_7.2.2_LIST_response_example            51.684k     45.809k i/s
        -    150.126k times in 2.904669s 3.277229s
                  utf8_in_list_mailbox            42.525k     36.422k i/s
        -    125.166k times in 2.943368s 3.436526s
                           xlist_inbox            43.099k     39.834k i/s
        -    129.726k times in 3.009982s 3.256683s

Comparison:
             list_with_various_flag_capitalizations
                    v0.4.5-8-g4611404d:     39942.7 i/s
                                 0.4.5:     35777.0 i/s - 1.12x  slower

                rfc3501_7.2.2_LIST_response_example
                    v0.4.5-8-g4611404d:     51684.4 i/s
                                 0.4.5:     45808.8 i/s - 1.13x  slower

                               utf8_in_list_mailbox
                    v0.4.5-8-g4611404d:     42524.8 i/s
                                 0.4.5:     36422.3 i/s - 1.17x  slower

                                        xlist_inbox
                    v0.4.5-8-g4611404d:     43098.6 i/s
                                 0.4.5:     39833.8 i/s - 1.08x  slower
```